refactor: move zipper to cevy::ecs::Query
