--- src/ecs/zipper_iterator.cpp
@@ -1,8 +1,0 @@
-/*
-** Agartha-Software, 2023
-** Cevy
-** File description:
-** Queries
-*/
-
-#include "zipper_iterator.hpp"
